change some values

+++ samples/f2.py
@@ -1,7 +1,7 @@
 def add(a,b):
     return a+b
 
-a =int(input("enter any no.= "))
-b =int(input("enter any no.= "))
+a =float(input("enter any no.= "))
+b =float(input("enter any no.= "))
 print(add(a,b))
 #print(a+b)
